app, testfixture: move App CR fixture to a shared package
Towards https://github.com/giantswarm/giantswarm/issues/5278
