+++ pkg/testfixture/app_cr.go
@@ -1,56 +1,55 @@
-package app
+package testfixture
 
 import (
 	"time"
 
-	"github.com/giantswarm/apiextensions/pkg/apis/application/v1alpha1"
 	applicationv1alpha1 "github.com/giantswarm/apiextensions/pkg/apis/application/v1alpha1"
-	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-func newTestAppCRFromFilled(modifyFunc func(*applicationv1alpha1.App)) *applicationv1alpha1.App {
+func NewAppCRFromFilled(modifyFunc func(*applicationv1alpha1.App)) *applicationv1alpha1.App {
 	appCR := &applicationv1alpha1.App{
 		TypeMeta: applicationv1alpha1.NewAppTypeMeta(),
-		ObjectMeta: v1.ObjectMeta{
+		ObjectMeta: metav1.ObjectMeta{
 			Name:      "test-app",
 			Namespace: "test-namespace",
 		},
-		Spec: v1alpha1.AppSpec{
+		Spec: applicationv1alpha1.AppSpec{
 			Catalog: "test-spec-catalog",
-			Config: v1alpha1.AppSpecConfig{
-				ConfigMap: v1alpha1.AppSpecConfigConfigMap{
+			Config: applicationv1alpha1.AppSpecConfig{
+				ConfigMap: applicationv1alpha1.AppSpecConfigConfigMap{
 					Name:      "test-spec-config-configmap-name",
 					Namespace: "test-spec-config-configmap-namespace",
 				},
-				Secret: v1alpha1.AppSpecConfigSecret{
+				Secret: applicationv1alpha1.AppSpecConfigSecret{
 					Name:      "test-spec-config-secret-name",
 					Namespace: "test-spec-config-secret-namespace",
 				},
 			},
-			KubeConfig: v1alpha1.AppSpecKubeConfig{
-				Secret: v1alpha1.AppSpecKubeConfigSecret{
+			KubeConfig: applicationv1alpha1.AppSpecKubeConfig{
+				Secret: applicationv1alpha1.AppSpecKubeConfigSecret{
 					Name:      "test-spec-kubeconfig-secret-name",
 					Namespace: "test-spec-kubeconfig-secret-namespace",
 				},
 			},
 			Name:      "test-spec-name",
 			Namespace: "test-spec-namespace",
-			UserConfig: v1alpha1.AppSpecConfig{
-				ConfigMap: v1alpha1.AppSpecConfigConfigMap{
+			UserConfig: applicationv1alpha1.AppSpecConfig{
+				ConfigMap: applicationv1alpha1.AppSpecConfigConfigMap{
 					Name:      "test-spec-userconfig-configmap-name",
 					Namespace: "test-spec-userconfig-configmap-namespace",
 				},
-				Secret: v1alpha1.AppSpecConfigSecret{
+				Secret: applicationv1alpha1.AppSpecConfigSecret{
 					Name:      "test-spec-userconfig-secret-name",
 					Namespace: "test-spec-userconfig-secret-namespace",
 				},
 			},
 			Version: "test-spec-version",
 		},
-		Status: v1alpha1.AppStatus{
+		Status: applicationv1alpha1.AppStatus{
 			AppVersion: "test-status",
-			Release: v1alpha1.AppStatusRelease{
-				LastDeployed: v1alpha1.DeepCopyTime{
+			Release: applicationv1alpha1.AppStatusRelease{
+				LastDeployed: applicationv1alpha1.DeepCopyTime{
 					Time: time.Date(2019, 2, 12, 12, 4, 0, 0, time.UTC),
 				},
 				Status: "test-status-release-status",
